% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bso-phenology-results.R
\name{analyse.phenology.bso}
\alias{analyse.phenology.bso}
\title{Analyse a BSO generated phenology}
\description{
Here, all functions are listed that are available to analyse the results of
a \code{\link[=bso_phenology]{bso_phenology()}} call.
}
\details{
Get BSO phenology properties:
\itemize{
\item \code{\link[=prop_dates]{prop_dates()}}
\item \code{\link[=prop_filial_generations]{prop_filial_generations()}}
\item \code{\link[=prop_first_date]{prop_first_date()}}
\item \code{\link[=prop_hatched_generations]{prop_hatched_generations()}}
\item \code{\link[=prop_last_date]{prop_last_date()}}
\item \code{\link[=prop_sister_broods]{prop_sister_broods()}}
\item \code{\link[=prop_stations]{prop_stations()}}
\item \code{\link[=prop_year]{prop_year()}}
}

Get BSO phenology results:
\itemize{
\item \code{\link[=bso_get_flight_df]{bso_get_flight_df()}}
\item \code{\link[=bso_get_flight_rst]{bso_get_flight_rst()}}
\item \code{\link[=bso_get_individuals_df]{bso_get_individuals_df()}}
\item \code{\link[=bso_get_individuals_rst]{bso_get_individuals_rst()}}
}

Plot BSO phenology results:
\itemize{
\item \code{\link[=bso_plot_flight_diagram]{bso_plot_flight_diagram()}}
\item \code{\link[=bso_plot_stage_diagram]{bso_plot_stage_diagram()}}
}
}
