% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrks-package.R
\docType{package}
\name{barrks-package}
\alias{barrks}
\alias{barrks-package}
\title{barrks: Calculate Bark Beetle Phenology Using Different Models}
\description{
Calculate the bark beetle phenology based on raster data or point-related data. There are models implemented for the bark beetle species Ips typographus and Pityogenes chalcographus. The models can be customized and their submodels (onset of infestation, beetle development, diapause initiation, mortality) can be combined. The following models are available in the package: PHENIPS-Clim (first-time release in this package), PHENIPS (Baier et al. 2007) \doi{10.1016/j.foreco.2007.05.020}, RITY (Ogris et al. 2019) \doi{10.1016/j.ecolmodel.2019.108775}, CHAPY (Ogris et al. 2020) \doi{10.1016/j.ecolmodel.2020.109137}, BSO (Jakoby et al. 2019) \doi{10.1111/gcb.14766}, Lange et al. (2006) \url{https://www.researchgate.net/publication/228511956_Thresholds_in_the_life_cycle_of_the_spruce_bark_beetle_under_climate_change}, Jönsson et al. (2011) \doi{10.1007/s10584-011-0038-4}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jjentschke.github.io/barrks/}
  \item \url{https://github.com/jjentschke/barrks}
  \item Report bugs at \url{https://github.com/jjentschke/barrks/issues}
}

}
\author{
\strong{Maintainer}: Jakob Jentschke \email{jakob.jentschke@forst.bwl.de}

Other contributors:
\itemize{
  \item FVA BW, Abt. Waldschutz \email{waldschutz.fva-bw@forst.bwl.de} [copyright holder, funder]
}

}
\keyword{internal}
