% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appearance.R
\name{barrks_colors}
\alias{barrks_colors}
\title{Get \code{barrks} default color palettes}
\usage{
barrks_colors(type = "raster")
}
\arguments{
\item{type}{Select the desired color palette. There are different variants
for particular purposes. Allowed values are \code{'raster'}, \code{'diagram_lines'},
\code{'diagram_fill'}, \code{'bso_flight'} and \code{'bso_stages'}.}
}
\value{
A character vector of hex colors.
}
\description{
Get \code{barrks} default color palettes.
}
\examples{
colors <- barrks_colors()

# use the colors of 'barrks' for your individual plot...
}
\seealso{
\code{\link[=barrks_labels]{barrks_labels()}}
}
