% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{barrks_data}
\alias{barrks_data}
\title{Load sample data}
\source{
https://opendata.dwd.de/climate_environment/CDC/grids_germany/daily/hyras_de/
https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/historical
https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/solar
}
\usage{
barrks_data(dataset = "raster")
}
\arguments{
\item{dataset}{Choose the dataset that should be returned.}
}
\description{
The package comes with sample data that allow the application of all models
available. The following datasets are available:
}
\details{
\itemize{
\item \code{raster} Contains a list of raster weather datasets for a sample area.
The data was taken from Deutscher Wetterdienst (DWD).
\item \code{stations} Contains sample station weather data for some cities in Germany.
The data was taken from Deutscher Wetterdienst (DWD).
Missing global radiation values were replaced by the mean value of the other stations.
}
}
