% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appearance.R
\name{barrks_labels}
\alias{barrks_labels}
\title{Get barrks default legend labels}
\usage{
barrks_labels(type = "raster")
}
\arguments{
\item{type}{Select the desired legend labels. There are different variants
for particular purposes. Allowed values are \code{'raster'}, \code{'diagram'},
\code{'bso_flight'} and \code{'bso_stages'}.}
}
\description{
Get barrks default legend labels
}
