% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bso-phenology.R
\name{bso_phenology}
\alias{bso_phenology}
\title{Calculate phenology with the BSO model}
\usage{
bso_phenology(
  .model = "bso",
  .data = NULL,
  .dates = NULL,
  .win = NULL,
  .ext = "tif",
  .onset = NULL,
  .diapause = NULL,
  .mortality = NULL,
  .submodels = c("onset", "diapause", "mortality", "development"),
  .setup_only = FALSE,
  .stations = NULL,
  .storage = NULL,
  .quiet = FALSE,
  ...
)
}
\arguments{
\item{.onset}{Note that it is not possible to pass an onset generate with
\code{\link[=create_onset]{create_onset()}} to \code{bso_phenology}. Only onsets from an other BSO phenology
(returned from \code{\link[=get_onset_rst]{get_onset_rst(..., as_doy = FALSE)}}) can be
used here.}
}
\value{
Returns a BSO phenology as a list. Look \link[=analyse.phenology.bso]{here} to find out how
a BSO phenology can be analysed. It is not recommended to access the list elements directly.
To be able to use the functions that are
available for phenology objects returned by \code{\link[=phenology]{phenology()}}, call \code{\link[=bso_translate_phenology]{bso_translate_phenology()}}.
}
\description{
As the BSO model works a bit different than the other models, a seperate
phenology function is implemented for this model. Note that while the
onset and the development submodels are needed to be taken from the BSO model,
the diapause and the mortality submodels are compatible with other models.
The usage of the function corresponds to \code{\link[=phenology]{phenology()}}. Look there for more
details.
}
