% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology-results.R
\name{categorize_generations_rst}
\alias{categorize_generations_rst}
\title{Make a numeric generations raster categorical}
\usage{
categorize_generations_rst(
  rst,
  colors = barrks_colors(),
  labels = barrks_labels()
)
}
\arguments{
\item{rst}{A numeric SpatRaster that represents bark beetle generations.
Sister broods are defined by adding \code{0.5} to the respective generation.}

\item{colors, labels}{Vectors of colors/labels starting from zero generations followed
consecutively by elements for the respective generations (including sister broods).}
}
\value{
A categorical SpatRaster.
}
\description{
Make a numeric generations raster categorical.
}
