% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-daylengths.R
\name{create_daylength_df}
\alias{create_daylength_df}
\title{Create a data frame of daylengths}
\usage{
create_daylength_df(lat, dates, .quiet = FALSE)
}
\arguments{
\item{lat}{(Named) vector of latitudes. The names of the vector indicate the
respective stations. If no names are provided, numbers are used instead.}

\item{dates}{Dates that should be processed.}

\item{.quiet}{If \code{TRUE}, messages are suppressed.}
}
\value{
A data frame with the columns \code{date}, \code{station} and \code{daylength}.
}
\description{
Generate a data frame of daylengths for given latitudes.
The package \link{geosphere} is required to use this function.
}
\seealso{
\code{\link[=create_daylength_rst]{create_daylength_rst()}}
}
