% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{list_models}
\alias{list_models}
\title{List all models}
\usage{
list_models()
}
\value{
A character vector.
}
\description{
Get the names of all available models.
}
\examples{
# print all available models
list_models()

}
