% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{load_phenology}
\alias{load_phenology}
\title{Load a phenology}
\usage{
load_phenology(
  .storage,
  .submodels = c("onset", "diapause", "mortality", "development"),
  .ext = ".tif",
  .quiet = FALSE
)
}
\arguments{
\item{.storage}{Path where the phenology should be loaded from.}

\item{.submodels}{Which submodels should be loaded.}

\item{.ext}{Extension of the raster files.}

\item{.quiet}{If \code{TRUE}, messages are suppressed.}
}
\description{
Loads a phenology from a path.
}
