% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bso.R
\encoding{UTF-8}
\name{model.bso.customize}
\alias{model.bso.customize}
\title{Customize the BSO model}
\arguments{
\item{dd_onset_start_date}{The date, when the degree days start to sum up ('MM-DD').}

\item{dd_onset_base}{Base temperature to calculate degree days to trigger the onset.}

\item{dd_onset_threshold}{Degree days that are required before the individuals
start regeneration feeding in \code{slot_dia} of the maturation stage. When the
regeneration feeding has finished, the onset is triggered.}

\item{slot_dia}{Maturation feeding slot where the individuals start regeneration feeding after
diapause.}

\item{k}{Factor for the calculation of the phloem temperature.}

\item{alpha, tlo, tup}{Parameters used to calculate the transition probabilities
for each stage (except preflight) in the following order: development, maturation feeding,
reproduction.}

\item{tfly_min, tfly_max, pfly_max, beta}{Parameters used to calculate the
transition probabilities for the preflight stage.}

\item{num_slots}{Named vector that defines the number of slots for each stage.
The development stage is subdivided into the stages \code{egg}, \code{larva} and \code{pupa}.}

\item{model_end_date}{Date when the model ends.}

\item{psis}{Probability that a sister brood will be established.}

\item{slot_sis}{Maturation feeding slot where the individuals start regeneration feeding before
they establish a sister brood.}

\item{diapause_first}{The day of year when the diapause could start at the
earliest.}

\item{diapause_last}{The day of year when the diapause could start at the
latest.}

\item{tdia_min}{The diapause will be initiated when the average daily
temperature falls below that value.}
}
\description{
This page describes the parameters
that can be used to customize the BSO model. The model was developed by
\insertCite{Jakoby2019;textual}{barrks}. Look \link[=model.bso.customize]{here} to find
out how to apply the model.
}
\details{
In \code{barrks}, \code{\link[=model]{model()}} is used to customize a model. The following code
illustrates which parameters are available for the BSO model and specifies their
default values.

\if{html}{\out{<div class="sourceCode">}}\preformatted{model('bso',

      dd_onset_start_date = '01-01',
      dd_onset_base = 5.124198,
      dd_onset_threshold = 100,
      slot_dia = 6,
      k = 0.02853738,
      c("alpha = 2.54906e-05", "alpha = 7.89e-05", "alpha = 1.00945e-05"),
      c("tlo = -12.97644", "tlo = 4.760089", "tlo = -4.424628"),
      c("tup = 36.0007", "tup = 40.02483", "tup = 39.9939"),
      tfly_min = 16.1064,
      tfly_max = 31.2901,
      pfly_max = 0.009863263,
      beta = 1.363763,
      c("num_slots = 11", "num_slots = 18", "num_slots = 45", "num_slots = 8", "num_slots = 8", "num_slots = 1"),
      model_end_date = '12-30',
      psis = 0.299445,
      slot_sis = 4,
      diapause_first = 210,
      diapause_last = 232,
      tdia_min = 16.45209
)
}\if{html}{\out{</div>}}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=model]{model()}}, \code{\link[=phenology]{phenology()}}, \code{\link{model.bso.apply}}

Other {model customizations}: 
\code{\link{model.chapy.customize}},
\code{\link{model.joensson.customize}},
\code{\link{model.lange.customize}},
\code{\link{model.phenips.customize}},
\code{\link{model.phenips_clim.customize}},
\code{\link{model.rity.customize}}
}
\concept{{model customizations}}
