% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-phenips.R
\encoding{UTF-8}
\name{model.phenips.customize}
\alias{model.phenips.customize}
\title{Customize PHENIPS}
\arguments{
\item{dd_onset_start_date}{The date, when the degree days start to sum up ('MM-DD').}

\item{dd_onset_base}{Base temperature to calculate degree days to trigger the onset.}

\item{dd_onset_threshold}{Degree days that are required to trigger the onset of
infestation. Additionally, the maximum temperature must exceed \code{tfly}.}

\item{tfly}{Minimum temperature that beetles need to fly.}

\item{dd_development_base}{Base temperature to calculate degree days for calculating
the beetles development.}

\item{dd_total_dev}{Degree days that are required for a generation to fully
develop}

\item{dev_start, dev_end}{Share in total development when the egg development starts and the
juvenile beetle's development ends respectively. Usable if the development
below/above these thresholds should account for mating, oviposition etc.}

\item{dev_sister_brood}{Share in the total development when a sister brood
will be established.}

\item{dev_mortal_min, dev_mortal_max}{Minimum/maximum share in the total
development of white stages (egg, larva, pupa). During these stages, the
beetles could die caused by a mortality event.}

\item{topt}{Temperature for optimal development.}

\item{tlow, tup}{Temperature below/above which no development happens.}

\item{func_btmean, func_btmax, func_btdiff}{Functions to caclulate the
effective bark temperature (see \insertCite{Baier2007;nobrackets}{barrks},
equations A.3 to A5). Each parameter will be passed as SpatRaster:
\itemize{
\item \code{tmean}: mean air temperature
\item \code{tmax}: maximum air temperature
\item \code{rad}: radiation
\item \code{btmax}: maximum bark temperature
}}

\item{daylength_dia}{When the daylength falls below this threshold, diapause
will be initiated.}

\item{model_end_date}{Date when the model ends and all white stages (egg, larva, pupa) die.}
}
\description{
This page describes the parameters
that can be used to customize PHENIPS. The model was developed by
\insertCite{Baier2007;textual}{barrks}. Look \link[=model.phenips.customize]{here} to find
out how to apply the model.
}
\details{
In \code{barrks}, \code{\link[=model]{model()}} is used to customize a model. The following code
illustrates which parameters are available for PHENIPS and specifies their
default values.

\if{html}{\out{<div class="sourceCode">}}\preformatted{model('chapy',

      dd_onset_start_date = '03-09',
      dd_onset_base = 7.4,
      dd_onset_threshold = 216.5,
      tfly = 15.6,
      dd_development_base = 7.4,
      dd_total_dev = 635.4,
      dev_start = 0,
      dev_end = 1,
      dev_sister_brood = 0.5,
      dev_mortal_min = ,
      dev_mortal_max = 0.8,
      func_ftmin = function (tmin) \{ 1.44 + 0.82 * tmin \},
      func_ftmean = function (tmean) \{ 0.5 + 0.81 * tmean \},
      func_ftmax = function (tmax) \{ 1.03 + 0.86 * tmax \},
      func_btmin = function (atmin) \{ 0.56 + 0.99 * atmin \},
      func_btmean = function (atmean) \{ -0.48 + 1.03 * atmean \},
      func_btmax = function (atmax) \{ 0.03 + 0.99 * atmax \},
      dt_low = 7.4,
      dt_up = 39.4,
      topt = 30,
      tmax = 41.97,
      alpha = 0.031,
      beta = 5.3,
      gamma = 1.25,
      daylength_dia = 13.6,
      model_end_date = '12-31'
)
}\if{html}{\out{</div>}}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=model]{model()}}, \code{\link[=phenology]{phenology()}}, \code{\link{model.phenips.apply}}

Other {model customizations}: 
\code{\link{model.bso.customize}},
\code{\link{model.chapy.customize}},
\code{\link{model.joensson.customize}},
\code{\link{model.lange.customize}},
\code{\link{model.phenips_clim.customize}},
\code{\link{model.rity.customize}}
}
\concept{{model customizations}}
