% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{params}
\alias{params}
\title{Get model parameters}
\usage{
params(m, ...)
}
\arguments{
\item{m}{Name of the model or the return value of another \code{model()}-call.}

\item{...}{List of parameters to customize the model.}
}
\value{
A list.
}
\description{
Get the parameters of a model.
}
\examples{
# print the first parameters of `phenips-clim`
head(params('phenips-clim'))
}
