% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-phenips-clim.R
\name{phenips_clim_get_dev_rates}
\alias{phenips_clim_get_dev_rates}
\title{Get development rates for PHENIPS-Clim}
\usage{
phenips_clim_get_dev_rates()
}
\value{
Returns a matrix that determines the development in relation to the
mean temperature and the temmperature amplitude. Col names indicate the mean
temperature and row names the temperature amplitude in °C * 10.
}
\description{
Get development rates for PHENIPS-Clim
}
