% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{save_phenology}
\alias{save_phenology}
\title{Save a phenology}
\usage{
save_phenology(
  pheno,
  .storage,
  .submodels = c("onset", "diapause", "mortality", "development"),
  .overwrite = FALSE,
  .ext = ".tif",
  .quiet = FALSE
)
}
\arguments{
\item{pheno}{A phenology, calculated with \code{\link[=phenology]{phenology()}}.}

\item{.storage}{Path that defines where to save the phenology.}

\item{.submodels}{Which submodels should be saved.}

\item{.overwrite}{Should the overwirte an existing storage?}

\item{.ext}{Extension for raster files.}

\item{.quiet}{If \code{TRUE}, messages are suppressed.}
}
\description{
Saves a phenology to a path.
}
