% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations_assign}
\alias{stations_assign}
\title{Assign stations to a phenology}
\usage{
stations_assign(pheno, stations)
}
\arguments{
\item{pheno}{A  phenology (see \code{\link[=phenology]{phenology()}})}

\item{stations}{Use the return value of \code{\link[=stations_create]{stations_create()}} here.}
}
\value{
Returns the phenology that was passed to the function with stations assigned.
}
\description{
Assign stations to a phenology.
}
\seealso{
\code{\link[=stations_create]{stations_create()}}
}
