% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations_cells}
\alias{stations_cells}
\title{Get the raster cells of stations}
\usage{
stations_cells(stations)
}
\arguments{
\item{stations}{Stations created with \code{\link[=stations_create]{stations_create()}} or obtained by
\code{\link[=prop_stations]{prop_stations()}}.}
}
\description{
Get the raster cells of stations.
}
