% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations_create}
\alias{stations_create}
\title{Build stations}
\usage{
stations_create(station_names, cells)
}
\arguments{
\item{station_names}{Character vector that specifies the names of the stations.}

\item{cells}{Numbers of the cells that should be represented by the stations.}
}
\value{
Returns stations that can be assigned to a phenology object or used for
\code{get_..._df}-functions.
}
\description{
Build stations that can be used to extract point data from a phenology.
}
\seealso{
\code{\link[=stations_assign]{stations_assign()}}
}
