% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-daylengths.R
\name{create_daylength_rst}
\alias{create_daylength_rst}
\title{Create daylength rasters}
\usage{
create_daylength_rst(
  template,
  dates = terra::time(template),
  crs = "EPSG:4258",
  .quiet = FALSE
)
}
\arguments{
\item{template}{(Multi-layer) SpatRaster that determines the
spatial extent of the result.}

\item{dates}{Dates that should be processed. If not specified, the dates of
\code{template} are used through \code{\link[terra:time]{terra::time()}}.}

\item{crs}{Coordinate reference system with longitude/latitude metrics.
It is used to project the raster coordinates to be able to retrieve the
latitude.}

\item{.quiet}{If \code{TRUE}, messages are suppressed.}
}
\value{
Returns a multi-layer SpatRaster. Each layer represents one date.
}
\description{
Generate a multi-layer SpatRaster of daylengths for a given template.
The package \code{geosphere} is required to use this function.
}
\seealso{
\code{\link[=create_daylength_df]{create_daylength_df()}}
}
