% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-lange.R
\encoding{UTF-8}
\name{model.lange.customize}
\alias{model.lange.customize}
\title{Customize the Lange model}
\arguments{
\item{dd_onset_start_date}{The date, when the degree days start to sum up ('MM-DD').}

\item{dd_onset_base}{Base temperature to calculate degree days to trigger the onset.}

\item{dd_onset_threshold}{Degree days that are required to trigger the onset of
infestation. Additionally, the maximum temperature must exceed \code{tfly}.}

\item{tfly}{Minimum temperature that beetles need to fly.}

\item{model_end_date}{Date when the model ends.}

\item{dd_base_stages}{Base temperatures to calculate degree days for the
different stages in the following order: egg, larva, pupa, juvenile adult.}

\item{dd_threshold_stages}{Thermal thresholds for the different stages in the
following order: egg, larva, pupa, juvenile adult.}

\item{first_lethal_date}{Date before which no mortality will be modeled.}

\item{tlethal}{Temperature threshold below which white stages (egg, larva, pupa) will die.}
}
\description{
This page describes the parameters
that can be used to customize Lange. The model was developed by
\insertCite{Lange2006;textual}{barrks}. Look \link[=model.lange.customize]{here} to find
out how to apply the model.
}
\details{
In \code{barrks}, \code{\link[=model]{model()}} is used to customize a model. The following code
illustrates which parameters are available for CHAPY and specifies their
default values.

\if{html}{\out{<div class="sourceCode">}}\preformatted{model('chapy',

      dd_onset_start_date = '03-09',
      dd_onset_base = 7.4,
      dd_onset_threshold = 216.5,
      tfly = 15.6,
      dd_development_base = 7.4,
      dd_total_dev = 635.4,
      dev_start = 0,
      dev_end = 1,
      dev_sister_brood = 0.5,
      dev_mortal_min = ,
      dev_mortal_max = 0.8,
      func_ftmin = function (tmin) \{ 1.44 + 0.82 * tmin \},
      func_ftmean = function (tmean) \{ 0.5 + 0.81 * tmean \},
      func_ftmax = function (tmax) \{ 1.03 + 0.86 * tmax \},
      func_btmin = function (atmin) \{ 0.56 + 0.99 * atmin \},
      func_btmean = function (atmean) \{ -0.48 + 1.03 * atmean \},
      func_btmax = function (atmax) \{ 0.03 + 0.99 * atmax \},
      dt_low = 7.4,
      dt_up = 39.4,
      topt = 30,
      tmax = 41.97,
      alpha = 0.031,
      beta = 5.3,
      gamma = 1.25,
      daylength_dia = 13.6,
      model_end_date = '12-31'
)
}\if{html}{\out{</div>}}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=model]{model()}}, \code{\link[=phenology]{phenology()}}, \code{\link{model.lange.apply}}

Other {model customizations}: 
\code{\link{model.bso.customize}},
\code{\link{model.chapy.customize}},
\code{\link{model.joensson.customize}},
\code{\link{model.phenips.customize}},
\code{\link{model.phenips_clim.customize}},
\code{\link{model.rity.customize}}
}
\concept{{model customizations}}
